Criada classe para tratamento de cada tipo de campo
