+++ src/Traits/CloudFirestoreDocumentResource.php
@@ -4,6 +4,8 @@ namespace Wead\Firestore\Traits;
 
 trait CloudFirestoreDocumentResource
 {
+    use CloudFirestoreDocumentFieldType;
+
     abstract public function getDocument($collection, $name);
 
     private function document($collection, $name, $createDefaultDoc = false)
@@ -47,146 +49,6 @@ trait CloudFirestoreDocumentResource
         $response->name = self::clearName($response->name);
 
         return $response;
-    }
-
-    public static function mapFieldValues($name, $fields)
-    {
-        $out = [
-            "name" => $name,
-            "fields" => [],
-        ];
-
-        $intField = function ($val) {
-            return ["integerValue" => trim($val)];
-        };
-
-        $stringField = function ($val) {
-            return ["stringValue" => (string) trim($val)];
-        };
-
-        $doubleField = function ($val) {
-            return ["doubleValue" => trim($val)];
-        };
-
-        $nullField = function ($val) {
-            return ["nullValue" => null];
-        };
-
-        $booleanField = function ($val) {
-            return ["booleanValue" => (bool) trim($val)];
-        };
-
-        $carbonTimestampField = function ($val) {
-            return ["timestampValue" => $val->toIso8601ZuluString()];
-        };
-
-        $mapField = function ($val, $testValue) use ($stringField, $intField, $doubleField, $nullField, $booleanField, &$arrayField, &$mapField) {
-            $result = [
-                "mapValue" => [
-                    "fields" => new \stdClass,
-                ],
-            ];
-
-            foreach ($val as $k => $v) {
-                if ($v) {
-                    $fnc = $testValue($v);
-
-                    $tmpValue = $$fnc($v, $testValue);
-
-                    if(is_array($tmpValue)) {
-                        if(isset($tmpValue['arrayValue'])) {
-                            $tmpValue = $tmpValue['arrayValue']['values'];
-                        } else if(isset($tmpValue['mapValue'])) {
-                            $tmpValue = $tmpValue['mapValue']['fields'];
-                        }
-
-                        if(!is_array($v)) {
-                            $result["mapValue"]["fields"]->{count((array) $result["mapValue"]["fields"])} = $v;
-                        } else {
-                            foreach($v as $fName => $fValue) {
-                                foreach($tmpValue as $tmpK => $tmpV) {
-                                    if(!is_array($tmpV)) {
-                                        $result["mapValue"]["fields"]->{$tmpK} = $v;
-                                    } else {
-                                        if(array_values($tmpV)[0] == $fValue) {
-                                            $result["mapValue"]["fields"]->{$fName} = $tmpV;
-                                            break;
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-
-            return $result;
-        };
-
-        $arrayField = function ($val, $testValue) use ($stringField, $intField, $doubleField, $nullField, $booleanField, &$mapField, &$arrayField) {
-            $result = [
-                "arrayValue" => [
-                    "values" => [],
-                ],
-            ];
-
-            foreach ($val as $k => $v) {
-                if ($v) {
-                    $fnc = $testValue($v);
-                    $result["arrayValue"]["values"][] = $$fnc($v, $testValue);
-                }
-            }
-
-            return sizeof($result["arrayValue"]["values"]) > 0 ? $result : false;
-        };
-
-        $testValue = function ($value) {
-            $result = null;
-
-            if (is_object($value)) {
-                if (!$value instanceof \Carbon\Carbon) {
-                    throw new \Exception("Unknown object type");
-                }
-
-                $result = "carbonTimestampField";
-            } else if (is_numeric($value) && (filter_var($value, FILTER_VALIDATE_INT, ['min_range' => 0]) != false || $value == "0" ) && substr_count($value, ".") == 0) {
-                $result = "intField";
-            } else if (is_numeric($value) && substr_count($value, ".") > 0 && preg_match('/[^0-9.]/', $value) == 0 && str_split($value)[0] != "0" && substr($value, -1) != "0") {
-                $result = "doubleField";
-            } else if (is_string($value)) {
-                $result = "stringField";
-            } else if (is_null($value)) {
-                $result = "nullField";
-            } else if (is_bool($value)) {
-                $result = "booleanField";
-            } else if (is_array($value)) {
-                $result = "arrayField";
-
-                foreach($value as $v) {
-                    if(is_array($v)) {
-                        $result = "mapField";
-                        break;
-                    }
-                }
-            }
-
-            if (!$result) {
-                throw new \Exception("Unknown value type to map");
-            }
-
-            return $result;
-        };
-
-        foreach ($fields as $field => $value) {
-            $fnc = $testValue($value);
-            $out["fields"][$field] = $$fnc($value, $testValue);
-
-            // if (!$out["fields"][$field]) {
-            //     unset($out["fields"][$field]);
-            // }
-        }
-
-        return $out;
     }
 
     private function readDocumentFields($doc)
